pyproject.toml
[project]
name = "c"
version = "0.1.0"
description = "An interesting http page"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pip>=25.1.1",
    "ruff>=0.11.13",
    "setuptools>=80.9.0",
]

[tool.ruff]
include = ["pyproject.toml", "src/**/*.py", "src/**/*.cgi"]
